Set up request formatter modules
